fixed map centering using insets

+++ HNMKMapView/libSrc/MKMapView+ZoomLevel.h
@@ -38,7 +38,7 @@
  *
  * \returns Current zoom level value
  */
-- (NSUInteger)zoomLevel;
+- (double)zoomLevel;
 
 - (NSNumber *)zoomLevelForFetch;
